gb10: cycle 55 step 12 — Phase D loader helpers + gamma dtype fix
Two related changes:

1. Loader helpers `tensor_to_bf16_bytes` + `f16_to_bf16` + `f32_to_bf16`
   + `fp8e4m3_to_bf16` (~80 LOC). Bf16 sibling of `tensor_to_f16_bytes`.
   The Gemma 4 checkpoint stores weights natively as bf16; this
   passes them through verbatim instead of narrowing. Foundation for
   Phase D wiring in cycle 55 step 13+ — once weight slots load as
   bf16, the bf16-native kernels added in steps 5-11 can consume
   them without runtime f16↔bf16 conversion.
   `#[allow(dead_code)]` until step 13+ wires it in.

2. Fix qk_rmsnorm_bf16 + qkv_rmsnorm_bf16 (cycle 55 steps 5 + 11)
   to keep gamma as `__half*`. The cycle-54 kernel
   `fused_norm_add_residual_bf16` already established the
   bf16-activation + f16-gamma convention; my new kernels were
   incorrectly making gamma bf16 too. Without this fix, switching
   dispatch to the bf16 kernels would have read f16 gamma bytes as
   bf16 → garbage. With this fix, dispatch can flip to the bf16
   kernels under the existing f16-gamma loader (no Phase D blocker
   for these specific kernels). Gamma read becomes
   `__half2float(gamma[i])` matching the cycle-54 norm-add convention.

Manifest sha256 entries refreshed (PTX content changed for both
qk_rmsnorm_bf16 and qkv_rmsnorm_bf16). Release build clean
(--features cuda,gb10); service restart smoke-test passes.

+++ v3/kernels/fused_qk_rmsnorm_bf16.cu
@@ -22,6 +22,7 @@
 // K output: [num_tokens, num_kv_heads, head_dim]  (can alias input)
 // q_gamma, k_gamma: [head_dim]
 
+#include <cuda_fp16.h>
 #include <cuda_bf16.h>
 
 extern "C"
@@ -30,8 +31,8 @@ __global__ void fused_qk_rmsnorm_bf16_kernel(
     const __nv_bfloat16* __restrict__ k_in,
     __nv_bfloat16* __restrict__ q_out,
     __nv_bfloat16* __restrict__ k_out,
-    const __nv_bfloat16* __restrict__ q_gamma,
-    const __nv_bfloat16* __restrict__ k_gamma,
+    const __half* __restrict__ q_gamma,
+    const __half* __restrict__ k_gamma,
     int num_tokens,
     int num_heads,
     int num_kv_heads,
@@ -50,7 +51,7 @@ __global__ void fused_qk_rmsnorm_bf16_kernel(
 
     const __nv_bfloat16* src = is_q ? q_in : k_in;
     __nv_bfloat16* dst = is_q ? q_out : k_out;
-    const __nv_bfloat16* gamma = is_q ? q_gamma : k_gamma;
+    const __half* gamma = is_q ? q_gamma : k_gamma;
 
     const int offset = (token * n_heads_this + head_local) * head_dim;
 
@@ -88,7 +89,7 @@ __global__ void fused_qk_rmsnorm_bf16_kernel(
     // Apply: out[i] = gamma[i] * x[i] * rms_inv
     for (int i = tid; i < head_dim; i += blockDim.x) {
         float v = __bfloat162float(src[offset + i]);
-        float g = __bfloat162float(gamma[i]);
+        float g = __half2float(gamma[i]);
         dst[offset + i] = __float2bfloat16(v * rms_inv * g);
     }
 }
